Implement better timing code
Better timing code has been implemented, at least as far a cycle counts go.  This change was made now due to the way paddle values are reported, which is directly tied to the number of cycles that have elapsed.  With this change, paddle values (mostly) report accurate numbers (although i can't get my joystick all the way to 255 - so something is still wrong).