Fix CPU problems:
1) invalid 6502 opcodes needed opcode length of 1 for disassembly purposes
2) A couple of the opcodes were missing opcode sizes (ADC specifically)
3) Fixed decimal mode on 65c02 to get the flags right.

decimal mode is really a pain.  The behavior is different (for flag) between 6502 and 65c02.  Comments in ADC command point to online refernece that I used to rework the ADC and SBC routines.  I'm not entirely sure that my overflow flag is accurate in these modes (based on functional testing), but I'm not sure that matters at least in the cases I have tested so far.